src/requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Werkzeug==3.1.3
python-dotenv==1.1.0
sirope==0.3.1
